+++ spec/firehose/account_message_spec.rb
@@ -68,6 +68,7 @@ describe Skyfall::Firehose::AccountMessage do
     it 'should have an operations field that returns []' do
       message = build_message(cbor)
       message.operations.should == []
+      message.ops.should == []
     end
 
     context 'for an active account' do