+++ spec/firehose/info_message_spec.rb
@@ -67,6 +67,7 @@ describe Skyfall::Firehose::InfoMessage do
     it 'should have an operations field that returns []' do
       message = build_message(cbor)
       message.operations.should == []
+      message.ops.should == []
     end
   end