+++ spec/firehose/labels_message_spec.rb
@@ -84,6 +84,7 @@ describe Skyfall::Firehose::LabelsMessage do
     it 'should have an operations field that returns []' do
       message = build_message(cbor)
       message.operations.should == []
+      message.ops.should == []
     end
   end