+++ spec/firehose/sync_message_spec.rb
@@ -79,6 +79,7 @@ describe Skyfall::Firehose::SyncMessage do
     it 'should have an operations field that returns []' do
       message = build_message(cbor)
       message.operations.should == []
+      message.ops.should == []
     end
   end
 end