+++ spec/firehose/unknown_message_spec.rb
@@ -44,6 +44,7 @@ describe Skyfall::Firehose::UnknownMessage do
     it 'should have an operations field that returns []' do
       message = build_message(cbor)
       message.operations.should == []
+      message.ops.should == []
     end
   end
 end