+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - un-deprecated `#path` in both `Operation` classes (marked deprecated in 0.7.0)
+- added `#ops` as alias for `#operations` in all message classes
 - fixed `CommitMessage#prev_data` throwing an error if `prevData` is nil
 - updated Firehose message constructors to verify that required fields aren't nil (not only that they're present), and for `LabelsMessage` that `labels` is an array